+++ src/components/index.ts
@@ -7,6 +7,7 @@
 
 export * from "./Accordion";
 export * from "./Alert";
+export * from "./AlertDialog";
 export * from "./Avatar";
 export * from "./Badge";
 export * from "./Breadcrumb";